Added the ability to override STS endpoint as well, and moved the resolver into util module
